Move eval callbacks to nvidia_nat_eval and guard core shim path
